SRTI v2 GUI v0.75.01 Update
Fixes bugs/features in SRTI v2 GUI:
1) Fix "initializeChannels" and "simulateChannels" to not print to config files if function is "".
2) Fix "endConditions" window to not state what stage it is active for: SRTI v2.00.00 Wrapper documentation states this is irrelevant, endConditions are checked at every stage.
3) Fix "endConditions" and "stageConditions" modal window not updating when opening for first time.
4) Show 'global' endConditions and stageConditions from project in any stage: gives user more context, and allows to delete from any stage.

For 4), it was accomplished by keeping "endConditions" and "stageConditions" attached to each stage object reference,  same as before, instead of to global simulator reference. This was to keep compatibility with .project files made with previous versions of the v2 GUI, but may change if this compatibility isn't deemed to be necessary.